[Refact] day view refactoring
